main/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
alembic==1.8.1
async-timeout==4.0.2
attrs==22.1.0
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.0
frozenlist==1.3.1
future==0.18.2
greenlet==1.1.3
gunicorn==20.1.0
idna==3.4
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.3.0
Mako==1.2.3
MarkupSafe==2.1.1
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pycparser==2.21
PyMySQL==1.0.2
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.41
tomli==2.0.1
urllib3==1.26.12
Werkzeug==2.2.2
yarl==1.8.1
